+++ crates/stdto_core/src/lib.rs
@@ -1,8 +1,8 @@
 mod enums;
 mod traits;
 
-pub use enums::HexMode;
-pub use traits::{ToBytes, ToHash, ToHex};
+pub use enums::{Endian, HexMode};
+pub use traits::{ToBytes, ToBytesOptions, ToHash, ToHex};
 
 pub mod error;