+++ crates/stdto_derive/src/lib.rs
@@ -1,6 +1,9 @@
-use proc_macro::TokenStream;
-use quote::quote;
-use syn::{parse_macro_input, DeriveInput};
+#[macro_use]
+mod cmn;
+use cmn::*;
+
+mod bytes;
+use bytes::*;
 
 #[proc_macro_attribute]
 pub fn serde(_: TokenStream, item: TokenStream) -> TokenStream {
@@ -18,7 +21,6 @@ pub fn to_bytes(input: TokenStream) -> TokenStream {
     let ast = parse_macro_input!(input as DeriveInput);
     let name = &ast.ident;
     let (impl_generics, ty_generics, where_clause) = &ast.generics.split_for_impl();
-
     quote! {
         impl #impl_generics stdto::ToBytes for #name #ty_generics #where_clause {
         }
@@ -27,19 +29,25 @@ pub fn to_bytes(input: TokenStream) -> TokenStream {
 }
 
 #[proc_macro_attribute]
-pub fn bytes(_: TokenStream, item: TokenStream) -> TokenStream {
+pub fn bytes(attr: TokenStream, item: TokenStream) -> TokenStream {
     let ast = parse_macro_input!(item as DeriveInput);
+    let attr = parse_macro_input!(attr as AttributeArgs);
+    let options = unwrap_error!(ToBytesOptions::try_from(attr)).to_expr();
+    let name = &ast.ident;
+    let (impl_generics, ty_generics, where_clause) = &ast.generics.split_for_impl();
     quote! {
         #[stdto::serde]
-        #[derive(stdto::ToBytes)]
         #ast
+        impl #impl_generics stdto::ToBytes for #name #ty_generics #where_clause {
+            const OPTIONS: stdto::ToBytesOptions = #options;
+        }
     }
     .into()
 }
 
 #[proc_macro_derive(ToHash)]
 pub fn to_hash(input: TokenStream) -> TokenStream {
-    let ast = parse_macro_input!(input as DeriveInput);
+    let ast = syn::parse_macro_input!(input as syn::DeriveInput);
     let name = &ast.ident;
     let (impl_generics, ty_generics, where_clause) = &ast.generics.split_for_impl();
 
@@ -52,7 +60,7 @@ pub fn to_hash(input: TokenStream) -> TokenStream {
 
 #[proc_macro_attribute]
 pub fn hash(_: TokenStream, item: TokenStream) -> TokenStream {
-    let ast = parse_macro_input!(item as DeriveInput);
+    let ast = syn::parse_macro_input!(item as syn::DeriveInput);
     quote! {
         #[derive(stdto::ToHash)]
         #ast