+++ src/lib.rs
@@ -1,5 +1,13 @@
 pub use stdto_core::error;
 
+// ----------------------------------------------------
+
+#[cfg(feature = "serde")]
+pub use stdto_core::serde;
+
+#[cfg(all(feature = "serde", feature = "derive"))]
+pub use stdto_derive::serde;
+
 // -----------------------------------------------------
 
 #[cfg(feature = "hex")]
@@ -8,12 +16,10 @@ pub use stdto_core::{HexMode, ToHex};
 // -----------------------------------------------------
 
 #[cfg(feature = "bytes")]
-pub use stdto_core::ToBytes;
+pub use stdto_core::{Endian, ToBytes, ToBytesOptions};
 
 #[cfg(all(feature = "bytes", feature = "derive"))]
-pub use stdto_core::serde;
-#[cfg(all(feature = "bytes", feature = "derive"))]
-pub use stdto_derive::{bytes, serde, ToBytes};
+pub use stdto_derive::{bytes, ToBytes};
 
 // -----------------------------------------------------