+++ crates/stdto_derive/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## [unreleased] - ReleaseDate
 
+## BREAKING CHANGES
+
+- Renamed the methods in `ToBytes` and Added the const ToBytesOptions.
+- Re-design [features].
+
+## Added
+
+- Added an attribute #[stdto::bytes(endian = "...")]. (little/big/native)
 
 ## [0.2.0] - 2022-12-10
 
@@ -22,8 +30,8 @@ Released by [@just-do-halee](https://github.com/just-do-halee).
 
 <!-- next-url -->
 
-[unreleased]: https://github.com/just-do-halee/stdto/compare/stdto_derive-v0.2.0...HEAD
+[unreleased]: https://github.com/just-do-halee/stdto/compare/v0.2.0...HEAD
 
-[0.2.0]: https://github.com/just-do-halee/stdto/compare/v0.1.0...stdto_derive-v0.2.0
+[0.2.0]: https://github.com/just-do-halee/stdto/compare/v0.1.0...v0.2.0
 
 [0.1.0]: https://github.com/just-do-halee/stdto/compare/v0.1.0...v0.1.0